Python-2/requirements.txt
python-telegram-bot==22.6
sqlalchemy==2.0.47
psycopg2-binary==2.9.11
flask==3.1.3
httpx==0.28.1
gunicorn==23.0.0
